pyproject.toml
[tool.poetry]
name = "web scraper"
version = "1.0"
description = "Simple Web Scraper"
authors = ["Your Name <you@example.com>"]
license = "None"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
beautifulsoup4 = "^4.12.0"
ipython = "^8.0.0"
openai = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
